+++ Business/Finance/StockBonds.php
@@ -4,25 +4,26 @@
  *
  * PHP Version 7.1
  *
- * @category   TBD
- * @package    TBD
+ * @package    phpOMS\Business\Finance
  * @copyright  Dennis Eichhorn
  * @license    OMS License 1.0
  * @version    1.0.0
- * @link       http://orange-management.com
+ * @link       http://website.orange-management.de
  */
-declare(strict_types=1);
+declare(strict_types = 1);
 
 namespace phpOMS\Business\Finance;
 
 /**
  * Finance class.
  *
- * @category   Log
- * @package    Framework
+ * @package    phpOMS\Business\Finance
  * @license    OMS License 1.0
- * @link       http://orange-management.com
+ * @link       http://website.orange-management.de
  * @since      1.0.0
+ * 
+ * @SuppressWarnings(PHPMD.CamelCaseParameterName)
+ * @SuppressWarnings(PHPMD.CamelCaseVariableName)
  */
 class StockBonds
 {
@@ -368,7 +369,6 @@ class StockBonds
      */
     public static function getZeroCouponBondEffectiveYield(float $F, float $PV, int $n) : float
     {
-        return pow($F / $PV, 1 / $n) - 1;
+        return (float) pow($F / $PV, 1 / $n) - 1;
     }
-
 }
