+++ Math/Stochastic/Distribution/ExponentialDistribution.php
@@ -4,24 +4,22 @@
  *
  * PHP Version 7.1
  *
- * @category   TBD
  * @package    TBD
  * @copyright  Dennis Eichhorn
  * @license    OMS License 1.0
  * @version    1.0.0
- * @link       http://orange-management.com
+ * @link       http://website.orange-management.de
  */
-declare(strict_types=1);
+declare(strict_types = 1);
 
 namespace phpOMS\Math\Stochastic\Distribution;
 
 /**
  * Exponential distribution.
  *
- * @category   Framework
- * @package    phpOMS\DataStorage\Database
+ * @package    Framework
  * @license    OMS License 1.0
- * @link       http://orange-management.com
+ * @link       http://website.orange-management.de
  * @since      1.0.0
  */
 class ExponentialDistribution
@@ -107,7 +105,7 @@ class ExponentialDistribution
      */
     public static function getVariance(float $lambda) : float
     {
-        return pow($lambda, -2);
+        return (float) pow($lambda, -2);
     }
 
     /**
