+++ Math/Stochastic/Distribution/PoissonDistribution.php
@@ -4,14 +4,13 @@
  *
  * PHP Version 7.1
  *
- * @category   TBD
  * @package    TBD
  * @copyright  Dennis Eichhorn
  * @license    OMS License 1.0
  * @version    1.0.0
- * @link       http://orange-management.com
+ * @link       http://website.orange-management.de
  */
-declare(strict_types=1);
+declare(strict_types = 1);
 
 namespace phpOMS\Math\Stochastic\Distribution;
 
@@ -20,10 +19,9 @@ use phpOMS\Math\Functions\Functions;
 /**
  * Well known functions class.
  *
- * @category   Framework
- * @package    phpOMS\DataStorage\Database
+ * @package    Framework
  * @license    OMS License 1.0
- * @link       http://orange-management.com
+ * @link       http://website.orange-management.de
  * @since      1.0.0
  */
 class PoissonDistribution
@@ -148,7 +146,7 @@ class PoissonDistribution
      */
     public static function getSkewness(float $lambda) : float
     {
-        return pow($lambda, -1 / 2);
+        return (float) pow($lambda, -1 / 2);
     }
 
     /**
@@ -162,7 +160,7 @@ class PoissonDistribution
      */
     public static function getFisherInformation(float $lambda) : float
     {
-        return pow($lambda, -1);
+        return (float) pow($lambda, -1);
     }
 
     /**
@@ -176,7 +174,7 @@ class PoissonDistribution
      */
     public static function getExKurtosis(float $lambda) : float
     {
-        return pow($lambda, -1);
+        return (float) pow($lambda, -1);
     }
 
     public static function getRandom()
