+++ Utils/StringCompare.php
@@ -4,14 +4,13 @@
  *
  * PHP Version 7.1
  *
- * @category   TBD
- * @package    TBD
+ * @package    phpOMS\Utils
  * @copyright  Dennis Eichhorn
  * @license    OMS License 1.0
  * @version    1.0.0
- * @link       http://orange-management.com
+ * @link       http://website.orange-management.de
  */
-declare(strict_types=1);
+declare(strict_types = 1);
 
 namespace phpOMS\Utils;
 
@@ -20,10 +19,9 @@ namespace phpOMS\Utils;
  *
  * This class helps to compare two strings
  *
- * @category   Framework
  * @package    phpOMS\Utils
  * @license    OMS License 1.0
- * @link       http://orange-management.com
+ * @link       http://website.orange-management.de
  * @since      1.0.0
  */
 class StringCompare
@@ -38,7 +36,7 @@ class StringCompare
 
     /**
      * Constructor.
-     * 
+     *
      * @param array $dictionary Dictionary
      *
      * @since  1.0.0
@@ -50,9 +48,9 @@ class StringCompare
 
     /**
      * Adds word to dictionary
-     * 
+     *
      * @param string $word Word to add to dictionary
-     * 
+     *
      * @return void
      *
      * @since  1.0.0
@@ -76,10 +74,10 @@ class StringCompare
         $bestScore = PHP_INT_MAX;
         $bestMatch = '';
 
-        foreach($this->dictionary as $word) {
+        foreach ($this->dictionary as $word) {
             $score = self::fuzzyMatch($word, $match);
 
-            if($score < $bestScore) {
+            if ($score < $bestScore) {
                 $bestScore = $score;
                 $bestMatch = $word;
             }
@@ -102,19 +100,19 @@ class StringCompare
     {
         $words1 = preg_split('/[ _-]/', $s1);
         $words2 = preg_split('/[ _-]/', $s2);
-        $total = 0;
+        $total  = 0;
 
-        foreach($words1 as $word1) {
+        foreach ($words1 as $word1) {
             $best = strlen($s2);
 
-            foreach($words2 as $word2) {
+            foreach ($words2 as $word2) {
                 $wordDist = levenshtein($word1, $word2);
 
-                if($wordDist < $best) {
+                if ($wordDist < $best) {
                     $best = $wordDist;
                 }
 
-                if($wordDist === 0) {
+                if ($wordDist === 0) {
                     break;
                 }
             }
@@ -152,25 +150,30 @@ class StringCompare
      */
     public static function valueLength(string $s1, string $s2) : int
     {
-        return abs(strlen($s1) - strlen($s2));
+        return (int) abs(strlen($s1) - strlen($s2));
     }
 
     /**
      * Calculate fuzzy match score.
      *
-     * @param string $s1 Word 1
-     * @param string $s2 Word 2
+     * @param string $s1          Word 1
+     * @param string $s2          Word 2
      * @param float $phraseWeight Weighting for phrase score
-     * @param float $wordWeight Weighting for word score
-     * @param float $minWeight Min weight
-     * @param float $maxWeight Max weight
+     * @param float $wordWeight   Weighting for word score
+     * @param float $minWeight    Min weight
+     * @param float $maxWeight    Max weight
      * @param float $lengthWeight Weighting for word length
      *
      * @return float
      *
      * @since  1.0.0
      */
-    public static function fuzzyMatch(string $s1, string $s2, float $phraseWeight = 0.5, float $wordWeight = 1, float $minWeight = 10, float $maxWeight = 1, float $lengthWeight = -0.3) : float
+    public static function fuzzyMatch(
+        string $s1, string $s2, 
+        float $phraseWeight = 0.5, float $wordWeight = 1, 
+        float $minWeight = 10, float $maxWeight = 1, 
+        float $lengthWeight = -0.3
+    ) : float
     {
         $phraseValue = self::valuePhrase($s1, $s2);
         $wordValue   = self::valueWords($s1, $s2);
@@ -179,5 +182,5 @@ class StringCompare
         return min($phraseValue * $phraseWeight, $wordValue * $wordWeight) * $minWeight
             + max($phraseValue * $phraseWeight, $wordValue * $wordWeight) * $maxWeight
             + $lengthValue * $lengthWeight;
-    } 
+    }
 }
